feat(cli): use react-scripts & create-react-app internally
