+++ cli/src/commands/test.js
@@ -1,31 +1,11 @@
-const path = require('path')
 const { reporter } = require('@dhis2/cli-helpers-engine')
-const { runCLI } = require('@jest/core')
-const fs = require('fs-extra')
 const exitOnCatch = require('../lib/exitOnCatch')
+const { craTest } = require('../lib/index.js')
 const loadEnvFiles = require('../lib/loadEnvFiles')
 const makePaths = require('../lib/paths')
 
-const getAppJestConfig = ({ jestConfigPath, paths }) => {
-    if (jestConfigPath) {
-        return require(path.resolve(paths.base, jestConfigPath))
-    } else if (fs.existsSync(paths.jestConfig)) {
-        return require(paths.jestConfig)
-    } else {
-        return {}
-    }
-}
-
-const handler = async ({
-    verbose,
-    cwd,
-    testRegex,
-    updateSnapshot,
-    coverage,
-    watch,
-    watchAll,
-    jestConfig: jestConfigPath,
-}) => {
+const handler = async ({ cwd }) => {
+    const testArgs = process.argv.slice(3)
     const paths = makePaths(cwd)
 
     const mode = (process.env.NODE_ENV = process.env.BABEL_ENV = 'test')
@@ -34,46 +14,7 @@ const handler = async ({
     reporter.info('Running tests...')
 
     await exitOnCatch(
-        async () => {
-            const defaultJestConfig = require(paths.jestConfigDefaults)
-            const appJestConfig = getAppJestConfig({
-                jestConfigPath,
-                paths,
-            })
-            const pkgJestConfig = require(paths.package).jest
-
-            const jestConfig = {
-                roots: ['./src'],
-                ...defaultJestConfig,
-                ...appJestConfig,
-                ...pkgJestConfig,
-            }
-
-            reporter.debug('Resolved jest config', jestConfig)
-
-            const ci = process.env.CI
-
-            const result = await runCLI(
-                {
-                    testPathPattern: testRegex,
-                    config: JSON.stringify(jestConfig),
-                    updateSnapshot: !ci && updateSnapshot,
-                    collectCoverage: coverage,
-                    watch: (!ci && watch) || undefined,
-                    watchAll: (!ci && watchAll) || undefined,
-                    ci,
-                    verbose: verbose,
-                },
-                [paths.base]
-            )
-
-            if (result.results.success) {
-                reporter.info(`Tests completed`)
-            } else {
-                reporter.error(`Tests failed`)
-                process.exit(1)
-            }
-        },
+        async () => await craTest({ cwd, testArgs }),
         {
             name: 'test',
             onError: () =>
