+++ cli/src/lib/i18n/index.js
@@ -1,9 +1,4 @@
-const extract = require('./extract')
-const generate = require('./generate')
-const validate = require('./validate')
-
-module.exports = {
-    extract,
-    generate,
-    validate,
-}
+module.exports.extract = require('./extract')
+module.exports.extractAndGenerate = require('./extract-and-generate.js')
+module.exports.generate = require('./generate')
+module.exports.validate = require('./validate')
