+++ cli/src/lib/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    ...require('./build/index.js'),
+    ...require('./init/index.js'),
+    ...require('./shell/index.js'),
+    ...require('./start/index.js'),
+}
