+++ cli/src/lib/init/index.js
@@ -1,0 +1,2 @@
+module.exports.initApp = require('./init-app.js')
+module.exports.initLibrary = require('./init-library.js')
