+++ cli/src/lib/start/index.js
@@ -1,0 +1,3 @@
+module.exports.createLocalProxyServer = require('./create-local-proxy-server.js')
+module.exports.notifyWhenPackageInvalid = require('./notify-when-package-invalid.js')
+module.exports.verifyIsApp = require('./verify-is-app.js')
